fix adress db for production and not deleted article another user

+++ controllers/articles.js
@@ -1,7 +1,6 @@
 const Article = require('../models/article');
 const NotFoundError = require('../errors/not-found-error');
 const BadRequestError = require('../errors/bad-request-error');
-const ForbiddenError = require('../errors/forbidden-error');
 
 module.exports.getArticles = (req, res, next) => {
   Article.find({})
@@ -26,12 +25,9 @@ module.exports.createArticle = (req, res, next) => {
 };
 
 module.exports.deleteArticle = (req, res, next) => {
-  Article.findOneAndDelete({ _id: req.params.articleId })
+  Article.findOneAndDelete({ _id: req.params.articleId, owner: req.user._id })
     .orFail()
     .then((article) => {
-      if (article.owner.toString() !== req.user._id) {
-        throw new ForbiddenError('Нельзя удалить чужую статью');
-      }
       res.status(200).send({ data: article });
     })
     .catch((err) => {
